If a test panics, shut down control plane and exit
Since run() is called by a goroutine, if the function panics somewhere,
execute() won't finish executing, and thus, the kubernetes control plane
won't get shut down. This adds a deferred function to the goroutine
started by startController that recovers from a panic so that it can
shut down the control plane before failing the test. It also exits to
prevent running other tests, because the control plane is shared between
tests and if it is in a bad state, other tests will inevitably have
errors.